syncval: Segregate submit/global operations
Queueu submit, semaphore, fence, swapchain, and wait operations
included.
