+++ BUILD.gn
@@ -247,6 +247,8 @@ vvl_sources = [
   "layers/sync/sync_model.h",
   "layers/sync/sync_renderpass.cpp",
   "layers/sync/sync_renderpass.h",
+  "layers/sync/sync_submit.cpp",
+  "layers/sync/sync_submit.h",
   "layers/sync/sync_sync_op.cpp",
   "layers/sync/sync_sync_op.h",
   "layers/sync/sync_utils.cpp",
